added lots of new Tower and Plot features

+++ db/migrate/20220511151303_create_plots.rb
@@ -1,9 +1,9 @@
 class CreatePlots < ActiveRecord::Migration[6.1]
   def change
     create_table :plots do |t|
-      t.integer :water_level
+      t.float :water_level
       t.integer :following_plot
-      t.integer :dehydration_rate
+      t.float :dehydration_rate
       t.integer :tower_id
       t.timestamps
     end
